+++ packages/terra-framework-docs/src/terra-dev-site/doc/embedded-content-consumer/example/CustomEventConsumer.jsx
@@ -15,11 +15,8 @@ const eventHandlers = [{
 const CustomEventConsumer = () => (
   <div id="CustomEvent">
     <EmbeddedContentConsumer
-      src="/terra-framework/#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-event-provider"
+      src="../../../#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-event-provider"
       title="Custom Event Example"
-      options={{
-        resizeConfig: { scrolling: true },
-      }}
       eventHandlers={eventHandlers}
     />
   </div>