+++ packages/terra-framework-docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 
 * Changed
   * Updated examples and tests for `terra-table` to consume updated onRowSelect callback.
+  * Updated XFC initialization for the Embedded Content Consumer examples.
 
 ## 1.49.0 - (December 1, 2023)