+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 ## Unreleased
 
+* Added
+  * Added documentation for exported constants of `terra-table` in the About page.
+
+* Changed
+  * Updated examples and tests for `terra-table` to consume updated onRowSelect callback.
+
+## 1.49.0 - (December 1, 2023)
+
+* Added
+  * Added new test to `terra-compact-interactive-list` for inconsistent width unit types in props.
+
+* Updated
+  * Updated examples, and tests to show the visual focus indicator on the iframe of the embedded content.
+  * Remove TODO comments that are no longer valid since relative paths work for examples, and tests for Embedded Content Consumers.
+
+## 1.48.0 - (November 29, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.47.0 - (November 23, 2023)
 
 * Added
@@ -24,7 +44,8 @@
   * Added `hasVisibleColumnHeaders` example for FlowsheetDataGrid.
   * Added examples and tests to `terra-data-grid` for sections in FlowsheetDataGrid.
   * Added documentation and examples for `terra-compact-interactive-list`.
-  
+  * Added a section on accessibility to `terra-embedded-content-consumer` implementation guide.
+
 * Updated
   * Updated About page of `terra-table` to provide accessibility documentation for Home and End keys.
   * Removed alpha notice for FlowsheetDataGrid.